A tool to save and load in chunks 1.21
In both 1.21 Java and Bedrock editions, you have the ability to save and load chunks of data. Additionally, you can specify a sublayer range to load, adjusting it to fit between existing chunks. Moreover, you have the option to include either blocks or entities as part of the loaded data. It may work with 1.18 and above, needs testing.